refactor: deduplicate 10 identical helper blocks from 3 Discord test files into shared discord_turn_fakes.py
Consolidate _FakeRest, _FakeGateway, _FakeOutboxManager, and 7 interaction
helper functions that were byte-for-byte identical across test_service_routing.py,
test_dispatch_validation.py, and test_flow_command_routing.py into the shared
tests/support/discord_turn_fakes.py module. Saves 550 net code lines (-0.10%).
